Store samples as int16_t internally instead of float
This dosen't change anything in terms of generated audio fidelity (the
float conversion is done at play-time instead of at load-time), but
allows the xm_context_t structure to be smaller and more compressible.

+++ src/xm_internal.h
@@ -99,7 +99,7 @@ struct xm_sample_s {
 	int8_t relative_note;
 	uint64_t latest_trigger;
 
-	float* data;
+	int16_t* data;
  };
  typedef struct xm_sample_s xm_sample_t;
